rep_orchestrator

Module event_loop

Top level event loop for processing kubernetes operations relating to test executions.

Handler functions

Each event type in the EventData enum needs to have an associated handler function written for it. These handlers are responsible for carrying out the required k8s operations for processing that event as well as updating the state of the parent TestExecution the event is for. As there are a number of different ways that we can end up unable to process an event, handler functions must be written to return a Result that the Event::handle method will use to record an Unrunnable status.

Structs

Claim

A claim for resolving a specific number of test executions.

Event

Raw event data paired with an associated TestExecution so we can track the status of the execution as we process it.

EventQueue

Coordinates queuing of k8s events to provide back pressure and prioritise running executions over newly submitted ones.

EventQueueState

Access to shared EventQueue state.

ProvisioningHandle

A handle for submitting provisioning requests to the EventQueue.

Enums

EventData

SubmitError

Errors that can be encountered when attempting to submit a test plan to the resolver task.

Constants

MSG_ARGO_WAIT

MSG_JOB_WAIT

Functions

event_loop_task

Run as a long lived task. This is an infinite loop that processes Events received on a channel that is shared with the axum server and the event loop’s own handler functions.